feat(deref): extracted various schema-types to separate files
