+++ packages/groqd/src/commands/as.ts
@@ -1,5 +1,5 @@
 import { GroqBuilder } from "../groq-builder";
-import { ExtractDocumentTypes } from "../types/schema-types";
+import { ExtractDocumentTypes } from "../types/document-types";
 
 declare module "../groq-builder" {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
