+++ packages/groqd/src/commands/notNull.ts
@@ -1,5 +1,5 @@
 import { GroqBuilder } from "../groq-builder";
-import { QueryConfig } from "../types/schema-types";
+import { QueryConfig } from "../types/query-config";
 import { ResultUtils } from "../types/result-types";
 import { Override } from "../types/utils";
 import { IGroqBuilderNotChainable } from "../types/public-types";
