+++ packages/groqd/src/index.ts
@@ -2,14 +2,14 @@
 import "./groq-builder";
 import "./commands";
 
-import type { QueryConfig } from "./types/schema-types";
+import type { QueryConfig } from "./types/query-config";
 import { GroqBuilder, GroqBuilderOptions, RootResult } from "./groq-builder";
 import { ZodMethods, zodMethods } from "./validation/zod";
 
 // Re-export all our public types:
 export * from "./groq-builder";
 export * from "./types/public-types";
-export * from "./types/schema-types";
+export * from "./types/query-config";
 export { zodMethods as zod } from "./validation/zod";
 export * from "./makeSafeQueryRunner";
 export * from "./validation/validation-errors";
@@ -73,3 +73,7 @@ export type GroqBuilderWithZod<TRootConfig extends QueryConfig> = ZodMethods &
   GroqBuilder<RootResult, TRootConfig>;
 
 export const createGroqBuilder = createGroqBuilderWithZod;
+export { ExtractRefType } from "./types/ref-types";
+export { RefType } from "./types/ref-types";
+export { ExtractDocumentTypes } from "./types/document-types";
+export { SchemaDocument } from "./types/document-types";
