+++ packages/groqd/src/types/document-types.ts
@@ -1,0 +1,6 @@
+export type SchemaDocument = { _type: string };
+
+export type ExtractDocumentTypes<TResultItem> = Extract<
+  TResultItem,
+  SchemaDocument
+>["_type"];
